requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
cryptography==44.0.2
    # via -r requirements.in
idna==3.10
    # via requests
pycparser==2.22
    # via cffi
pyjwt==2.10.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
urllib3==2.3.0
    # via requests
